refactor : Data dto 분리
