+++ data/src/main/java/com/example/data/dto/response/OpeningHoursDto.kt
@@ -1,0 +1,13 @@
+package com.example.data.dto.response
+
+import com.example.domain.model.OpeningHours
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class OpeningHoursDto(
+    val open: TimeInfoData,
+    val close: TimeInfoData,
+)
+
+internal fun OpeningHoursDto.toDomainModel(): OpeningHours =
+    OpeningHours(open.toDomainModel(), close.toDomainModel())
